completed create directories
