+++ src/components/App/App.js
@@ -1,18 +1,18 @@
-import agent from '../agent';
-import Header from './Header';
 import React from 'react';
 import { connect } from 'react-redux';
-import { APP_LOAD, REDIRECT } from '../constants/actionTypes';
+import agent from '../../agent';
+import Header from '../Header/Header';
+import { APP_LOAD, REDIRECT } from '../../constants/actionTypes';
 import { Route, Switch } from 'react-router-dom';
-import Article from '../components/Article';
-import Editor from '../components/Editor';
-import Home from '../components/Home';
-import Login from '../components/Login';
-import Profile from '../components/Profile';
-import ProfileFavorites from '../components/ProfileFavorites';
-import Register from '../components/Register';
-import Settings from '../components/Settings';
-import { store } from '../store';
+import Article from '../Pages/Article';
+import Editor from '../Editor/Editor';
+import Home from '../Pages/Home';
+import Login from '../Login/Login';
+import Profile from '../Profile/Profile';
+import ProfileFavorites from '../ProfileFavorites/ProfileFavorites';
+import Register from '../Register/Register';
+import Settings from '../Settings/Settings';
+import { store } from '../../store';
 import { push } from 'connected-react-router';
 
 const mapStateToProps = state => {
