+++ src/components/ArticleList/ArticleList.js
@@ -1,5 +1,5 @@
-import ArticlePreview from './ArticlePreview';
-import ListPagination from './ListPagination';
+import ArticlePreview from '../ArticlePreview/ArticlePreview';
+import ListPagination from '../ListPagination/ListPagination';
 import React from 'react';
 
 const ArticleList = props => {
