+++ src/components/ListPagination/ListPagination.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import agent from '../agent';
+import agent from '../../agent';
 import { connect } from 'react-redux';
-import { SET_PAGE } from '../constants/actionTypes';
+import { SET_PAGE } from '../../constants/actionTypes';
 
 const mapDispatchToProps = dispatch => ({
   onSetPage: (page, payload) =>
