+++ src/components/Pages/Article/CommentInput.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import agent from '../../agent';
+import agent from '../../../agent';
 import { connect } from 'react-redux';
-import { ADD_COMMENT } from '../../constants/actionTypes';
+import { ADD_COMMENT } from '../../../constants/actionTypes';
 
 const mapDispatchToProps = dispatch => ({
   onSubmit: payload =>
