+++ src/components/Pages/Article/DeleteButton.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import agent from '../../agent';
+import agent from '../../../agent';
 import { connect } from 'react-redux';
-import { DELETE_COMMENT } from '../../constants/actionTypes';
+import { DELETE_COMMENT } from '../../../constants/actionTypes';
 
 const mapDispatchToProps = dispatch => ({
   onClick: (payload, commentId) =>
