+++ src/components/Pages/Home/MainView.js
@@ -1,8 +1,8 @@
-import ArticleList from '../ArticleList';
+import ArticleList from '../../ArticleList/ArticleList';
 import React from 'react';
-import agent from '../../agent';
+import agent from '../../../agent';
 import { connect } from 'react-redux';
-import { CHANGE_TAB } from '../../constants/actionTypes';
+import { CHANGE_TAB } from '../../../constants/actionTypes';
 
 const YourFeedTab = props => {
   if (props.token) {
