+++ src/components/Pages/Home/index.js
@@ -2,13 +2,13 @@ import Banner from './Banner';
 import MainView from './MainView';
 import React from 'react';
 import Tags from './Tags';
-import agent from '../../agent';
+import agent from '../../../agent';
 import { connect } from 'react-redux';
 import {
   HOME_PAGE_LOADED,
   HOME_PAGE_UNLOADED,
   APPLY_TAG_FILTER
-} from '../../constants/actionTypes';
+} from '../../../constants/actionTypes';
 
 // YOU WILL DELETE NEXT LINE SOON
 import ToDelete from './ToDelete';
