+++ src/components/Profile/Profile.js
@@ -1,14 +1,14 @@
-import ArticleList from './ArticleList';
+import ArticleList from '../ArticleList/ArticleList';
 import React from 'react';
 import { Link } from 'react-router-dom';
-import agent from '../agent';
+import agent from '../../agent';
 import { connect } from 'react-redux';
 import {
   FOLLOW_USER,
   UNFOLLOW_USER,
   PROFILE_PAGE_LOADED,
   PROFILE_PAGE_UNLOADED
-} from '../constants/actionTypes';
+} from '../../constants/actionTypes';
 
 const EditProfileSettings = props => {
   if (props.isUser) {
