+++ src/components/ProfileFavorites/ProfileFavorites.js
@@ -1,12 +1,12 @@
-import { Profile, mapStateToProps } from './Profile';
+import { Profile, mapStateToProps } from '../Profile/Profile';
 import React from 'react';
 import { Link } from 'react-router-dom';
-import agent from '../agent';
+import agent from '../../agent';
 import { connect } from 'react-redux';
 import {
   PROFILE_PAGE_LOADED,
   PROFILE_PAGE_UNLOADED
-} from '../constants/actionTypes';
+} from '../../constants/actionTypes';
 
 const mapDispatchToProps = dispatch => ({
   onLoad: (pager, payload) =>
