+++ src/components/Settings/Settings.js
@@ -1,12 +1,12 @@
-import ListErrors from './ListErrors';
+import ListErrors from '../ListErrors/ListErrors';
 import React from 'react';
-import agent from '../agent';
+import agent from '../../agent';
 import { connect } from 'react-redux';
 import {
   SETTINGS_SAVED,
   SETTINGS_PAGE_UNLOADED,
   LOGOUT
-} from '../constants/actionTypes';
+} from '../../constants/actionTypes';
 
 class SettingsForm extends React.Component {
   constructor() {
